requirements.txt
flask
pymysql
bcrypt
boto3
python-dotenv
werkzeug

